Updated, consolidated, edited, and such...
I've tried to update, consolidate, and edit the README because there were many 404 links and extra sections that I thought could be combined. I've left all direct quotes and personal things but we still need to fix a few of the links as they are 404's now and I'm not sure where to find them.